Added Autofac dependency.
