+++ DotNet/Web/LayeredAppDemo/LayeredApp.Web/Controllers/Api/Messages/Customer/CustomerViewModel.cs
@@ -1,0 +1,10 @@
+﻿
+namespace LayeredApp.Web.Controllers.Api.Messages.Customer
+{
+    public class CustomerViewModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Address { get; set; }
+    }
+}
