+++ DotNet/Web/LayeredAppDemo/LayeredApp.Web/Controllers/Api/Messages/Customer/GetCustomersResponse.cs
@@ -1,0 +1,8 @@
+﻿
+namespace LayeredApp.Web.Controllers.Api.Messages.Customer
+{
+    public class GetCustomersResponse : ResponseMessageBase
+    {
+        public CustomerViewModel[] Customers { get; set; }
+    }
+}
