+++ DotNet/Web/LayeredAppDemo/LayeredApp.Web/Controllers/Api/Messages/ResponseMessageBase.cs
@@ -1,0 +1,19 @@
+﻿
+using System;
+
+namespace LayeredApp.Web.Controllers.Api.Messages
+{
+    public class ResponseMessageBase
+    {
+        public bool Valid { get; set; }
+        public string Message { get; set; }
+        public string[] Errors { get; set; }
+
+        public ResponseMessageBase()
+        {
+            Valid = false;
+            Message = String.Empty;
+            Errors = new string[] {};
+        }
+    }
+}
