+++ DotNet/Web/LayeredAppDemo/LayeredApp.Web/Startup.cs
@@ -9,7 +9,7 @@ namespace LayeredApp.Web
     {
         public void Configuration(IAppBuilder app)
         {
-            //ConfigureAuth(app);            
+            ConfigureAuth(app);            
         }
     }
 }
