+++ DotNet/Web/LayeredAppDemo/LayeredApp.Web/packages.config
@@ -1,6 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Antlr" version="3.4.1.9004" targetFramework="net452" />
+  <package id="Autofac" version="3.5.2" targetFramework="net452" />
+  <package id="Autofac.WebApi2" version="3.4.0" targetFramework="net452" />
   <package id="bootstrap" version="3.0.0" targetFramework="net452" />
   <package id="EntityFramework" version="6.1.1" targetFramework="net452" />
   <package id="jQuery" version="1.10.2" targetFramework="net452" />
